feat: Managing active GoalHolders in GoalMaker
Need to make goals "unscorable" on after they've been used with a ball.
First, I need to communicate to the GoalMaker that the Goal needs to
update to some "unscorable state". Actually, I think that's it lol. This
requires adding an event listener to the GoalMaker in MakeGoal() after
it makes the goal.